Revert "Changed the Zero checks from BigDecimal to BigInteger to ignore the"
This reverts commit f7c3dc8c69073abea12719e7db771e50b3149d41.

+++ src/main/java/io/company/brewcraft/model/TaxRate.java
@@ -1,7 +1,6 @@
 package io.company.brewcraft.model;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.math.RoundingMode;
 
 import javax.persistence.Column;
@@ -47,7 +46,7 @@ public class TaxRate extends BaseEntity {
 
     @JsonIgnore
     public boolean isSet() {
-        return this.value != null && !this.value.toBigInteger().equals(BigInteger.ZERO);
+        return this.value != null && !this.value.equals(BigDecimal.ZERO);
     }
 
     public static boolean isSet(TaxRate taxRate) {
